+++ frontend/android/app/src/main/java/io/github/langxi13/babydiary/MainActivity.java
@@ -1,12 +1,5 @@
 package io.github.langxi13.babydiary;
 
-import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        registerPlugin(NativeShareReceiverPlugin.class);
-        super.onCreate(savedInstanceState);
-    }
-}
+public class MainActivity extends BridgeActivity {}